+++ src/vs/workbench/contrib/output/browser/outputServices.ts
@@ -16,9 +16,10 @@ import { ITextModelService, ITextModelContentProvider } from 'vs/editor/common/s
 import { ITextModel } from 'vs/editor/common/model';
 import { ILogService } from 'vs/platform/log/common/log';
 import { ILifecycleService } from 'vs/workbench/services/lifecycle/common/lifecycle';
-import { IOutputChannelModel, IOutputChannelModelService } from 'vs/workbench/contrib/output/common/outputChannelModel';
+import { IOutputChannelModel } from 'vs/workbench/contrib/output/common/outputChannelModel';
 import { IViewsService } from 'vs/workbench/common/views';
 import { OutputViewPane } from 'vs/workbench/contrib/output/browser/outputView';
+import { IOutputChannelModelService } from 'vs/workbench/contrib/output/common/outputChannelModelService';
 
 const OUTPUT_ACTIVE_CHANNEL_KEY = 'output.activechannel';
