+++ src/vs/workbench/contrib/output/common/outputChannelModelService.ts
@@ -3,13 +3,47 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IOutputChannelModelService, AbstractOutputChannelModelService } from 'vs/workbench/contrib/output/common/outputChannelModel';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
-import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
+import { createDecorator, IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IFileService } from 'vs/platform/files/common/files';
 import { toLocalISOString } from 'vs/base/common/date';
 import { dirname, joinPath } from 'vs/base/common/resources';
+import { DelegatedOutputChannelModel, FileOutputChannelModel, IOutputChannelModel } from 'vs/workbench/contrib/output/common/outputChannelModel';
+import { URI } from 'vs/base/common/uri';
+
+export const IOutputChannelModelService = createDecorator<IOutputChannelModelService>('outputChannelModelService');
+
+export interface IOutputChannelModelService {
+	readonly _serviceBrand: undefined;
+
+	createOutputChannelModel(id: string, modelUri: URI, mimeType: string, file?: URI): IOutputChannelModel;
+
+}
+
+export abstract class AbstractOutputChannelModelService {
+
+	declare readonly _serviceBrand: undefined;
+
+	constructor(
+		private readonly outputLocation: URI,
+		@IFileService protected readonly fileService: IFileService,
+		@IInstantiationService protected readonly instantiationService: IInstantiationService
+	) { }
+
+	createOutputChannelModel(id: string, modelUri: URI, mimeType: string, file?: URI): IOutputChannelModel {
+		return file ? this.instantiationService.createInstance(FileOutputChannelModel, modelUri, mimeType, file) : this.instantiationService.createInstance(DelegatedOutputChannelModel, id, modelUri, mimeType, this.outputDir);
+	}
+
+	private _outputDir: Promise<URI> | null = null;
+	private get outputDir(): Promise<URI> {
+		if (!this._outputDir) {
+			this._outputDir = this.fileService.createFolder(this.outputLocation).then(() => this.outputLocation);
+		}
+		return this._outputDir;
+	}
+
+}
 
 export class OutputChannelModelService extends AbstractOutputChannelModelService implements IOutputChannelModelService {
